Split directory.cpp into unix/win32
