+++ libmediation/fs/osdep/directory_win32.cpp
@@ -1,64 +1,34 @@
-
-#include "directory.h"
-
-#include <errno.h>
-#include <sys/stat.h>
-#include <sys/types.h>
-
-#include <sstream>
-
-#include "file.h"
-
-#ifdef _WIN32
 #include <windows.h>
-#else
-#include <dirent.h>
-#include <memory.h>
-#include <stddef.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <unistd.h>
-#endif
+
+#include "../directory.h"
+#include "../directory_priv.h"
+
+#include "../file.h"
 
 using namespace std;
 
 char getDirSeparator()
 {
-#ifdef _WIN32
     return '\\';
-#else
-    return '/';
-#endif
 }
 
 string extractFileDir(const string& fileName)
 {
-    size_t index = fileName.find_last_of('/');
+    size_t index = fileName.find_last_of('\\');
     if (index != string::npos)
         return fileName.substr(0, index + 1);
-#ifdef _WIN32
-    index = fileName.find_last_of('\\');
-    if (index != string::npos)
-        return fileName.substr(0, index + 1);
-#endif
 
     return "";
 }
 
 bool fileExists(const string& fileName)
 {
-#ifdef _WIN32
     File f;
     return f.open(fileName.c_str(), File::ofRead | File::ofOpenExisting);
-#else
-    struct stat buf;
-    return stat(fileName.c_str(), &buf) == 0;
-#endif
 }
 
 uint64_t getFileSize(const std::string& fileName)
 {
-#ifdef _WIN32
     File f;
     if (f.open(fileName.c_str(), File::ofRead | File::ofOpenExisting))
     {
@@ -69,11 +39,6 @@ uint64_t getFileSize(const std::string& fileName)
     {
         return 0;
     }
-#else
-    struct stat fileStat;
-    auto res = stat(fileName.c_str(), &fileStat) == 0;
-    return res ? static_cast<uint64_t>(fileStat.st_size) : 0;
-#endif
 }
 
 bool createDir(const std::string& dirName, bool createParentDirs)
@@ -91,13 +56,6 @@ bool createDir(const std::string& dirName, bool createParentDirs)
             if (dirEnd != string::npos)
             {
                 string parentDir = dirName.substr(0, dirEnd);
-#if __linux__ == 1 || (defined(__APPLE__) && defined(__MACH__))
-                if (mkdir(parentDir.c_str(), S_IREAD | S_IWRITE | S_IEXEC) == -1)
-                {
-                    if (errno != EEXIST)
-                        return false;
-                }
-#elif defined(_WIN32)
                 if (parentDir.size() == 0 || parentDir[parentDir.size() - 1] == ':' || parentDir == string("\\\\.") ||
                     parentDir == string("\\\\.\\") ||                                                // UNC patch prefix
                     (strStartWith(parentDir, "\\\\.\\") && parentDir[parentDir.size() - 1] == '}'))  // UNC patch prefix
@@ -107,23 +65,15 @@ bool createDir(const std::string& dirName, bool createParentDirs)
                     if (GetLastError() != ERROR_ALREADY_EXISTS)
                         return false;
                 }
-#endif
             }
         }
     }
 
-#if __linux__ == 1 || (defined(__APPLE__) && defined(__MACH__))
-    return mkdir(dirName.c_str(), S_IREAD | S_IWRITE | S_IEXEC) == 0;
-#elif defined(_WIN32)
     return CreateDirectory(toWide(dirName).data(), 0) != 0;
-#endif
 }
 
 bool deleteFile(const string& fileName)
 {
-#if __linux__ == 1 || (defined(__APPLE__) && defined(__MACH__))
-    return unlink(fileName.c_str()) == 0;
-#else
     if (DeleteFile(toWide(fileName).data()))
     {
         return true;
@@ -135,14 +85,7 @@ bool deleteFile(const string& fileName)
     }
 
     return DeleteFile(toWide(fileName).data()) != 0;
-#endif
 }
-
-//-----------------------------------------------------------------------------
-#if defined(_WIN32)
-/** windows implementation */
-//-----------------------------------------------------------------------------
-#include <windows.h>
 
 bool findFiles(const string& path, const string& fileMask, vector<string>* fileList, bool savePaths)
 {
@@ -194,79 +137,9 @@ bool findDirs(const string& path, vector<string>* dirsList)
 
     return true;
 }
-#elif __linux__ == 1 || (defined(__APPLE__) && defined(__MACH__))
-#include <fnmatch.h>
-
-bool findFiles(const string& path, const string& fileMask, vector<string>* fileList, bool savePaths)
-{
-    dirent** namelist;
-
-    int n = scandir(path.c_str(), &namelist, 0, 0);  // alphasort);
-
-    if (n < 0)
-    {
-        return false;
-    }
-    else
-    {
-        while (n--)
-        {
-            if (namelist[n]->d_type == DT_REG)
-            {
-                string fileName(namelist[n]->d_name);
-
-                if (0 == fnmatch(fileMask.c_str(), fileName.c_str(), 0))
-                    fileList->push_back(savePaths ? path + fileName : fileName);
-            }
-            free(namelist[n]);
-        }
-        free(namelist);
-    }
-
-    return true;
-}
-
-bool findDirs(const string& path, vector<string>* dirsList)
-{
-    dirent** namelist;
-
-    int n = scandir(path.c_str(), &namelist, 0, 0);  // alphasort);
-
-    if (n < 0)
-    {
-        return false;
-    }
-    else
-    {
-        while (n--)
-        {
-            if (namelist[n]->d_type == DT_DIR)
-            {
-                string dirName(namelist[n]->d_name);
-
-                // we save only normal ones
-                if ("." != dirName && ".." != dirName)
-                    dirsList->push_back(path + dirName + "/");
-            }
-            free(namelist[n]);
-        }
-        free(namelist);
-    }
-
-    return true;
-}
-#endif
-//-----------------------------------------------------------------------------
 
 bool findFilesRecursive(const string& path, const string& mask, vector<string>* const fileList)
 {
-    // finding files
-    findFiles(path, mask, fileList, true);
-
-    vector<string> dirList;
-    findDirs(path, &dirList);
-
-    for (unsigned int d_idx = 0; d_idx != dirList.size(); d_idx++) findFilesRecursive(dirList[d_idx], mask, fileList);
-
+    recurseDirectorySearch(findFiles, findDirs, path, mask, fileList);
     return true;
 }
